creating user session token on-demand
