+++ packages/commercetools/api-client/src/index.ts
@@ -27,6 +27,7 @@ import { Config, ConfigurableConfig } from './types/setup';
 export * from './types/Api';
 export * from './types/setup';
 export * as cartActions from './helpers/cart/actions';
+export * from './helpers/token';
 
 let apolloClient: ApolloClient<any> = null;
